file-manager: refactor to prepare for reuse in repo browser
