+++ app/assets/javascripts/Components/markus_file_manager.jsx
@@ -1,91 +1,12 @@
+/* MarkUs-specific customization of react-keyed-file-browser library.
+ * Provides customized versions of the components in that library.
+ */
 import React from 'react'
-import { render } from 'react-dom'
 import ClassNames from 'classnames'
 import HTML5Backend, { NativeTypes } from 'react-dnd-html5-backend'
 import { DragDropContext, DragSource, DropTarget } from 'react-dnd'
 
 import { RawFileBrowser, Headers, FileRenderers, BaseFileConnectors } from 'react-keyed-file-browser'
-
-class StudentFileManger extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      files: []
-    };
-  }
-
-  componentDidMount() {
-    window.modal_addnew = new ModalMarkus('#addnew_dialog');
-    this.fetchData();
-  }
-
-  fetchData = () => {
-    fetch(Routes.populate_student_file_manager_assignment_submissions_path(this.props.assignment_id), {
-      credentials: 'same-origin',
-      headers: {
-        'content-type': 'application/json'
-      }
-    }).then(data => data.json())
-      .then(data => this.setState({files: data}));
-  };
-
-  handleCreateFiles = (files, prefix) => {
-    let data = new FormData();
-    files.forEach(f => data.append('new_files[]', f, f.name));
-    data.append('path', '/' + prefix); // Server expects path with leading slash (TODO: fix that)
-    $.post({
-      url: Routes.update_files_assignment_submissions_path(this.props.assignment_id),
-      data: data,
-      processData: false,  // tell jQuery not to process the data
-      contentType: false   // tell jQuery not to set contentType
-    }).then(this.fetchData);
-  };
-
-  handleDeleteFile = (fileKey) => {
-    let deleteFiles = [];
-    this.state.files.map((file) => {
-      if (file.key === fileKey) {
-        deleteFiles.push(file)
-      }
-    });
-    if (!deleteFiles) {
-      return;
-    }
-
-    let file = deleteFiles[0];
-    let file_revisions = {};
-    file_revisions[file.key] = file.last_modified_revision;
-    $.post({
-      url: Routes.update_files_assignment_submissions_path(this.props.assignment_id),
-      data: {
-        delete_files: [file.key],
-        file_revisions: file_revisions
-      }
-    }).then(this.fetchData)
-      .then(this.endAction);
-  };
-
-  handleActionBarDeleteClick = (event) => {
-    event.preventDefault();
-    if (this.state.selection) {
-      this.handleDeleteFile(this.state.selection);
-    }
-  };
-
-  render() {
-    return (
-      <FileManager
-        files={this.state.files}
-        // Override TableHeader
-        headerRenderer={FileManagerHeader}
-        fileRenderer={FileManagerFile}
-
-        onDeleteFile={this.handleDeleteFile}
-        onCreateFiles={this.handleCreateFiles}
-      />
-    );
-  }
-}
 
 
 class RawFileManager extends RawFileBrowser {
@@ -96,6 +17,10 @@ class RawFileManager extends RawFileBrowser {
   };
 
   renderActionBar(selectedItem) {
+    if (this.props.readOnly) {
+      return <div className="item-actions">&nbsp;</div>;
+    }
+
     const selectionIsFolder = (selectedItem && !selectedItem.size)
     let filter
     if (this.props.canFilter) {
@@ -264,8 +189,6 @@ class RawFileManager extends RawFileBrowser {
   }
 }
 
-let FileManager = DragDropContext(HTML5Backend)(RawFileManager);
-
 
 class FileManagerHeader extends Headers.TableHeader {
   render() {
@@ -398,13 +321,24 @@ class FileManagerFile extends FileRenderers.RawTableFile {
 }
 
 FileManagerFile = (
-    DragSource('file', BaseFileConnectors.dragSource, BaseFileConnectors.dragCollect)(
+  DragSource('file', BaseFileConnectors.dragSource, BaseFileConnectors.dragCollect)(
     DropTarget(['file', 'folder', NativeTypes.FILE],
-      BaseFileConnectors.targetSource,
-      BaseFileConnectors.targetCollect,
+               BaseFileConnectors.targetSource,
+               BaseFileConnectors.targetCollect,
     )(FileManagerFile))
-  );
+);
 
-export function makeStudentFileManager(elem, props) {
-  render(<StudentFileManger {...props} />, elem);
+
+let FileManager = DragDropContext(HTML5Backend)(RawFileManager);
+
+FileManager.defaultProps = {
+  headerRenderer: FileManagerHeader,
+  fileRenderer: FileManagerFile
+};
+
+
+export default FileManager;
+export {
+  FileManagerHeader,
+  FileManagerFile
 }
