+++ app/javascript/packs/application.js
@@ -20,7 +20,7 @@ import { makeStudentTable } from 'javascripts/Components/student_table';
 import { makeAssignmentSummaryTable } from 'javascripts/Components/assignment_summary_table';
 import { makeExamScanLogTable } from 'javascripts/Components/exam_scan_log_table';
 import { makeMarksSpreadsheet } from 'javascripts/Components/marks_spreadsheet';
-import { makeStudentFileManager } from 'javascripts/Components/student_file_manager';
+import { makeSubmissionFileManager } from 'javascripts/Components/submission_file_manager';
 
 import 'javascripts/react_config';
 
@@ -33,4 +33,4 @@ window.makeTATable = makeTATable;
 window.makeAssignmentSummaryTable = makeAssignmentSummaryTable;
 window.makeExamScanLogTable = makeExamScanLogTable;
 window.makeMarksSpreadsheet = makeMarksSpreadsheet;
-window.makeStudentFileManager = makeStudentFileManager;
+window.makeSubmissionFileManager = makeSubmissionFileManager;
