+++ app/views/submissions/file_manager.html.erb
@@ -7,8 +7,20 @@
 
   <script>
     document.addEventListener('DOMContentLoaded', () => {
-      makeStudentFileManager(document.getElementById('file_manager'),
-                             {assignment_id: <%= @assignment.id %>});
+      makeSubmissionFileManager(document.getElementById('file_manager'),
+        {
+          assignment_id: <%= @assignment.id %>,
+          readOnly: <%= !@assignment.allow_web_submits.to_s %>
+        });
+
+      window.setInterval(() => {
+        $.get({
+          url: '<%= server_time_assignment_submissions_path(@assignment.id) %>',
+          success: function(data) {
+            document.getElementById('current_time').innerHTML = data;
+          }
+        });
+      }, 60000);
     });
   </script>
 <% end %>
@@ -18,44 +30,12 @@
           short_identifier: sanitize(@assignment.short_identifier)) %>
 <% end %>
 
-<div>
+<% content_for :additional_headings do %>
   <strong><%= t('server_time') %></strong>
-  <span id='current_time'>
-    <%= l(Time.zone.now)%>
-  </span>
+  <span id='current_time'><%= l(Time.zone.now)%></span>
+<% end %>
 
-  <script>
-    window.setInterval(function () {
-      $.ajax({
-        url: '<%= server_time_assignment_submissions_path(@assignment.id) %>',
-        method: 'GET',
-        success: function(data) {
-          document.getElementById('current_time').innerHTML = data;
-        }
-      });
-    }, 60000);
-    </script>
-</div>
-
-<div class='wrapper'>
-  <% if @assignment.vcs_submit %>
-    <p>
-      <%= t('student.submission.current_path') %>
-      <%= link_to '[root]', action: 'file_manager', path: '/' %>
-      <%
-        path_memory = '/'
-        @path.split('/').each do |folder|
-          if (folder != '')
-            path_memory = File.join(path_memory, folder) %>
-            <%= link_to(" / #{folder}", action: 'file_manager', path: path_memory).html_safe %>
-          <% end
-        end
-      %>
-    </p>
-  <% end %>
-
-  <div id='file_manager'></div>
-</div>
+<div id='file_manager'></div>
 
 <aside class='dialog' id='addnew_dialog'>
   <h2><%= t('add_new') %></h2>
