+++ app/views/submissions/repo_browser.html.erb
@@ -84,7 +84,7 @@
       <% end %>
     </h2>
 
-    <p>
+    <p>=
       <%= t('browse_submissions.current_path') %>
       <%
         path_memory = '/'
